Created MainFrameGUI, Updated multiple GUIs and Model classes.
Added MainFrameGUI for better reusability and access for createButtonEvents() and createGUI() so CurrentGUI has less code and better readability. Updated Cart and Inventory so that userID is the correct type (int now instead of String). Updated User so that each instance also has a Cart and Inventory. Made Cart, Inventory and Product Serializable as well.